+++ dbt_schemas/README.md
@@ -1,0 +1,7 @@
+## Schemas and their generator Pydantic models
+
+Use this command to generate the models:
+
+```bash
+$ datamodel-codegen --input dbt_schemas/dbt_manifest_v9_schema.json --input-file-type jsonschema --output dbt_schemas/dbt_manifest_v9.py
+```
