--- schemas/README.md
@@ -1,7 +1,0 @@
-## Schemas and their generator Pydantic models
-
-Use this command to generate the models:
-
-```bash
-$ datamodel-codegen --input schemas/dbt_manifest_v9_schema.json --input-file-type jsonschema --output schemas/dbt_manifest_v9.py
-```
